Implement some memory and typecheck relevant APIs.
1. implement dvmAllocObject.
2. implement typecheck relevant APIs.
3. fixed some build errors.

+++ porting/src/opl_file.c
@@ -217,7 +217,7 @@ int32_t file_open(const uint16_t* name, int32_t nameLen, int32_t mode, int32_t *
     else /* Success */
     {
         DVMTraceInf("CPL_file_open: _open %s %d returned 0x%x",
-            getAsciiName(name, nameLen), mode, res);
+            getAsciiName(name, nameLen), mode, ((int32_t)res));
 
         *handle = (int32_t)res;
